student_frolova_ilona/src/main/java/lv/javaguru/travel/insurance/core/underwriting/calculators/medical/CountryDefaultDayRateCalculator.java
package lv.javaguru.travel.insurance.core.underwriting.calculators.medical;

import lombok.RequiredArgsConstructor;
import lv.javaguru.travel.insurance.core.domain.CountryDefaultDayRate;
import lv.javaguru.travel.insurance.core.repositories.CountryDefaultDayRateRepository;
import lv.javaguru.travel.insurance.dto.TravelCalculatePremiumRequest;
import org.springframework.stereotype.Component;

import java.math.BigDecimal;

@Component
@RequiredArgsConstructor
public class CountryDefaultDayRateCalculator {

    private final CountryDefaultDayRateRepository countryRepository;

    public BigDecimal calculate(TravelCalculatePremiumRequest request) {
        return countryRepository.findByCountryIc(request.getCountry())
                .map(CountryDefaultDayRate::getDefaultDayRate)
                .orElseThrow(() -> new RuntimeException(
                        "Country day rate not found by country id = " + request.getCountry())
                );
    }
}
